% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_d3_pdf.R
\name{save_d3_pdf}
\alias{save_d3_pdf}
\title{Save a D3 visualization as a PDF file}
\usage{
save_d3_pdf(d3, file, background = "white", delay = 0.5)
}
\arguments{
\item{d3}{D3 visualization to save.}

\item{file}{File to save PDF into.}

\item{background}{Text string giving the html background color of}

\item{delay}{Delay (in seconds) to wait before saving the SVG.
the widget. Defaults to white.}
}
\description{
Save a D3 visualization as a PDF file
}
\examples{
library(r2d3)

viz <- r2d3(
  data = c(0.3, 0.6, 0.8, 0.95, 0.40, 0.20),
  script = system.file("examples/barchart.js", package = "r2d3")
)

save_d3_pdf(viz, file = tempfile(fileext = ".pdf"))
}
