% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_d3_svg.R
\name{save_d3_svg}
\alias{save_d3_svg}
\title{Save a D3 visualization as a SVG file}
\usage{
save_d3_svg(d3, file, background = "white", delay = 0.5)
}
\arguments{
\item{d3}{D3 visualization to save.}

\item{file}{File to save SCG into.}

\item{background}{Text string giving the html background color of
the widget. Defaults to white.}

\item{delay}{Delay (in seconds) to wait before saving the SVG.}
}
\description{
Save a D3 visualization as a SVG file
}
\examples{
library(r2d3)

viz <- r2d3(
  data = c(0.3, 0.6, 0.8, 0.95, 0.40, 0.20),
  script = system.file("examples/barchart.js", package = "r2d3")
)

save_d3_svg(viz, file = tempfile(fileext = ".svg"))
}
